Fix for md5 hashing bug in login and registration process
Created a new md5 instance each time a password or token needs to be hashed. This prevents any possibility of carrying over data from previous hashing operations, ensuring correct results. The bug was originally identified when the user was unable to login after logging out due to a password mismatch.